(#47)refactor: MemberRepository 위치 수정했습니다.
