+++ src/main/java/com/coverflow/global/jwt/service/JwtService.java
@@ -3,8 +3,8 @@ package com.coverflow.global.jwt.service;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.coverflow.member.domain.Member;
-import com.coverflow.member.domain.MemberRepository;
 import com.coverflow.member.domain.Role;
+import com.coverflow.member.infrastructure.MemberRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.Getter;
