+++ src/main/java/com/coverflow/global/oauth2/service/CustomOAuth2UserService.java
@@ -4,8 +4,8 @@ import com.coverflow.global.oauth2.CustomOAuth2User;
 import com.coverflow.global.oauth2.OAuthAttributes;
 import com.coverflow.global.util.NicknameUtil;
 import com.coverflow.member.domain.Member;
-import com.coverflow.member.domain.MemberRepository;
 import com.coverflow.member.domain.SocialType;
+import com.coverflow.member.infrastructure.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
