+++ src/main/java/com/coverflow/global/util/NicknameUtil.java
@@ -1,7 +1,7 @@
 package com.coverflow.global.util;
 
 import com.coverflow.member.domain.Member;
-import com.coverflow.member.domain.MemberRepository;
+import com.coverflow.member.infrastructure.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
